+++ src/core/Text.cpp
@@ -1,21 +1,19 @@
 /* Text.cpp */
 
-#include <cstdlib>
 #include <cstring>
 #include <core/Text.h>
 
 Core::Text::Text() : m_pos_x(0), m_pos_y(0), m_color(0), m_has_changed(true),
-                     m_length(0), m_str(NULL) {
+        m_length(0), m_str(NULL) {
 
 }
 
-Core::Text::Text(int8_t pos_x, int8_t pos_y) : m_pos_x(pos_x), m_pos_y(pos_y),
-                                               m_color(0), m_has_changed(true),
-                                               m_length(0), m_str(NULL) {
+Core::Text::Text(int pos_x, int pos_y) : m_pos_x(pos_x), m_pos_y(pos_y),
+        m_color(0), m_has_changed(true), m_length(0), m_str(NULL) {
 
 }
 
-Core::Text::Text(int8_t pos_x, int8_t pos_y, int8_t color, const char *str) :
+Core::Text::Text(int pos_x, int pos_y, int color, const char *str) :
         m_pos_x(pos_x), m_pos_y(pos_y), m_color(color), m_has_changed(true),
         m_length(0), m_str(NULL) {
     update_text(str);
@@ -30,7 +28,7 @@ void Core::Text::update_text(const char *new_str) {
     }
 }
 
-void Core::Text::update_pos(int8_t pos_x, int8_t pos_y) {
+void Core::Text::update_pos(int pos_x, int pos_y) {
     if (m_pos_x != pos_x || m_pos_y != pos_y) {
         m_pos_x = pos_x;
         m_pos_y = pos_y;
