+++ src/include/core/ChooseMode.h
@@ -12,11 +12,11 @@ namespace Core {
         public:
             ChooseMode();
             void init();
-            void update(int8_t *state, int8_t *mode);
+            void update(int *state, int *mode);
 
         protected:
             bool m_selected;
-            int8_t m_mode;
+            int m_mode;
             uint8_t m_DASup, m_DASdown;
             ::Text m_modes_strings[NB_MODES];
     };
