+++ src/include/core/Digits.h
@@ -10,12 +10,12 @@ namespace Core {
         public:
             Digits();
 
-            void init(int8_t pos_x, int8_t pos_y);
+            void init(int pos_x, int pos_y);
             void update(uint32_t number);
 
         protected:
             uint32_t m_value;
-            int8_t m_pos_x, m_pos_y;
+            int m_pos_x, m_pos_y;
             bool m_has_changed;
     };
 }
