+++ src/include/core/LineClearParticles.h
@@ -3,13 +3,11 @@
 #ifndef CORE_LINE_CLEAR_PARTICLES_H
 #define CORE_LINE_CLEAR_PARTICLES_H
 
-#include <stdint.h>
-
 namespace Core {
     class LineClearParticles {
         public:
             LineClearParticles();
-            void setEmitter(uint8_t position_x, uint8_t position_y);
+            void setEmitter(int position_x, int position_y);
             void update();
             void init();
             bool end();
