+++ src/include/core/Timer.h
@@ -11,8 +11,8 @@ namespace Core {
             uint32_t m_frames;
             float m_time;
 
-            int8_t m_pos_x, m_pos_y;
-            int8_t str[6];
+            int m_pos_x, m_pos_y;
+            uint8_t str[6];
 
             void init();
             void start();
